requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.30.2
tokenizers==0.13.3
huggingface_hub==0.14.1
chromadb==0.4.24
openai==1.30.1
anthropic==0.25.7
torch==2.2.2
torchvision==0.17.2
Pillow==10.0.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
